+++ lib/disposable_cached_images.dart
@@ -1,16 +1,22 @@
 library disposable_cached_images;
 
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:http/http.dart' as http;
 
 import './cache/interface.dart';
-import './image_info_data/image_info_data.dart';
+import './image_info_data.dart';
+import './image_size/size_export.dart';
 
-part './disposable_image_widget.dart';
-part './globals.dart';
-part './image_provider/image_provider.dart';
-part './image_provider/image_provider_arguments.dart';
-part './image_provider/image_provider_state.dart';
-part './image_provider/interface.dart';
-part './image_provider/used_image_provider.dart';
+part './image_providers/assets_image.dart';
+part './image_providers/dynamic_size_image.dart';
+part './image_providers/image_provider_arguments.dart';
+part './image_providers/image_provider_state.dart';
+part './image_providers/interface.dart';
+part './image_providers/network_image.dart';
+part './image_providers/used_image_provider.dart';
+part './run_app_wrapper.dart';
+part './widget/helpers.dart';
+part './widget/main_widget.dart';