+++ lib/image_size/size_export.dart
@@ -1,0 +1,3 @@
+export 'size_stub.dart'
+    if (dart.library.io) './size.dart'
+    if (dart.library.html) './web_image_size.dart';